day4/requirements.txt
streamlit>=1.43.0
google-genai>=1.68.0
fastembed>=0.8.0
faiss-cpu>=1.10.0
pdfplumber>=0.11.0
networkx>=3.4.0
pyvis>=0.3.2
python-dotenv>=1.0.0
numpy
requests
